implement TDATA message type
This commit implements the TDATA message type, which is the first composite
message type to be implemented in pyigtl. The message is abstracted in
an object-oriented fashion by composition, i.e. TDataMessage requires a
list of TDataRecord objects in the constructor.

So far, this implementation has only been tested with the project's unit testing
framework. It still needs to be tested with the reference implementation of
OpenIGTLink.